Client chat message packet moved into apply
