+++ src/main/kotlin/packets/PacketHandler.kt
@@ -657,77 +657,6 @@ class PacketHandler(
     }
 
     /**
-     * Handles when a chat message is received
-     */
-    @PacketReceiver
-    fun onChatMessage(packet: ClientChatMessagePacket) {
-        val message = packet.message
-
-        if (message.length > 255) {
-            client.player.sendMessage(
-                Component.text("Message is too long")
-                    .color(NamedTextColor.RED)
-            )
-
-            return
-        }
-
-        if (message.startsWith('/') && message.length > 1) {
-            val command = message.substring(1)
-            val commandSource = client.player
-
-            @Suppress("TooGenericExceptionCaught")
-            val result: Int = try {
-                CommandManager.dispatcher.execute(command, commandSource)
-            } catch (e: CommandSyntaxException) {
-                CommandCodes.ILLEGAL_SYNTAX.id
-            } catch (e: Exception) {
-                Bullet.logger.warn("Error running command `$message`:", e)
-                return
-            }
-
-            if (result == CommandCodes.SUCCESS.id) return
-
-            when (result) {
-                CommandCodes.UNKNOWN.id ->
-                    commandSource.sendMessage(
-                        Component.text("Unknown command")
-                            .color(NamedTextColor.RED)
-                    )
-
-                CommandCodes.ILLEGAL_ARGUMENT.id,
-                CommandCodes.ILLEGAL_SYNTAX.id ->
-                    commandSource.sendMessage(
-                        Component.text("Invalid command syntax, try typing /help")
-                            .color(NamedTextColor.RED)
-                    )
-
-                CommandCodes.INVALID_PERMISSIONS.id ->
-                    commandSource.sendMessage(
-                        Component.text("You don't have permission to use this command")
-                            .color(NamedTextColor.RED)
-                    )
-            }
-            return
-        }
-
-        val formattedMessage = message.replace('&', '§')
-
-        val event = PlayerChatEvent(client.player, formattedMessage)
-        EventManager.fire(event)
-        if (event.isCancelled) return
-
-        val textComponent = Component.text()
-            .append(Component.text().content("<").color(NamedTextColor.GRAY))
-            .append(Component.text().content(client.player.username).color(TextColor.color(0x55FFFF)))
-            .append(Component.text().content("> ").color(NamedTextColor.GRAY))
-            .append(MiniMessage.miniMessage().deserialize(formattedMessage))
-            .build()
-
-        Bullet.broadcast(textComponent)
-    }
-
-    /**
      * Handles when the client responds to the server keep alive packet to tell the server the client is still online
      * It also calculates the round trip time (RTT) and updates the players ping
      */
